Improve sound effect generation
The old implementation had several limitations:

Firstly, when a pattern still had some time remaining, it filled the complete buffer.
This typically meant that sounds lasted a bit longer which was especially noticeable for short patterns.

Secondly, when no more patterns remained, the entire buffer would be cleared via a memset.
This would typically wipe samples from the last pattern which still had to be played.
This was especially noticeable after increasing the buffer size and/or decreasing the sample rate.
For the default buffer size and sample rate this effect was minimal, as during a single update period most of the buffer would be emptied (as 44100 / 25 ~= 2048).

Thirdly, the implementation could be made more efficient, in particular the inner loops that generate the samples.